CMake: improve dependency to python
most of the targets do not need to include
PYTHON_INCLUDE_DIRS, because they don't have a direct
dependency. This was only necessary because
FreeCADBase didn't populate it's public
include directories with the use of target_include_directories

mark PYTHON_INCLUDE_DIRS as SYSTEM headers, to
hide compiler warnings on Ubuntu 18.04 with python2.